pyproject.toml
[project]
name = "draw_apriltags"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Aaron Marburg", email = "amarburg@uw.edu"},
]
dependencies = ["pillow>=11.2.1", "fpdf2>=2.8.3", "numpy>=2.0.2"]
requires-python = "==3.9.*"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
make-apriltag-sheet = "draw_apriltags.make_apriltag_sheet:entrypoint"


[tool.pdm]
distribution = true
